pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sre-gym"
version = "1.0.0"
description = "SRE incident response RL environment for OpenEnv"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "httpx"]

[tool.hatch.build.targets.wheel]
packages = [
  "server",
  "client.py",
  "models.py",
  "__init__.py"
]
